#41 stored-object をドメイン集約からインフラ層のデータクラスへ降格
StoredObjectMetadata は R2 メタデータの D1 キャッシュ + ダウンロード
カウンターであり、DDD 集約としてのドメイン振る舞いが不足していた。

- domain/stored-object/ → infra/stored-object/ へ移動
- IPersisted/IUnpersisted ジェネリクスと IEntity インターフェースを除去
- equals()/toJSON() メソッドを除去（不要な DDD パターン）
- リポジトリインターフェースのファイル命名を統一
  (`-` 区切り → `.` 区切り)
