+++ app/backend/handlers/api/files/index.test.ts
@@ -4,7 +4,7 @@ import { describe, expect, it, vi } from "vitest";
 import { ContentType } from "../../../domain/shared/content-type.vo";
 import { ETag } from "../../../domain/shared/etag.vo";
 import { ObjectKey } from "../../../domain/shared/object-key.vo";
-import { StoredObjectMetadata } from "../../../domain/stored-object/stored-object-metadata.entity";
+import { StoredObjectMetadata } from "../../../infra/stored-object/stored-object-metadata";
 import { filesApp } from ".";
 
 // Mock drizzle
