+++ app/backend/infra/stored-object/stored-object-metadata.command-repository.interface.ts
@@ -1,0 +1,11 @@
+import type { StoredObjectMetadata } from "./stored-object-metadata";
+import type { ObjectKey } from "../../domain/shared/object-key.vo";
+
+export interface IStoredObjectMetadataCommandRepository {
+  upsert(
+    metadata: StoredObjectMetadata,
+    options?: { preserveDownloadCount?: boolean },
+  ): Promise<StoredObjectMetadata>;
+  deleteByObjectKey(objectKey: ObjectKey): Promise<void>;
+  incrementDownloadCount(objectKey: ObjectKey): Promise<void>;
+}
