+++ app/backend/infra/stored-object/stored-object-metadata.query-repository.interface.ts
@@ -1,0 +1,9 @@
+import type { StoredObjectMetadata } from "./stored-object-metadata";
+import type { ObjectKey } from "../../domain/shared/object-key.vo";
+
+export interface IStoredObjectMetadataQueryRepository {
+  findAll(): Promise<readonly StoredObjectMetadata[]>;
+  findByObjectKey(
+    objectKey: ObjectKey,
+  ): Promise<StoredObjectMetadata | undefined>;
+}
