+++ app/backend/services/sync.service.test.ts
@@ -3,14 +3,14 @@ import { describe, expect, it, vi } from "vitest";
 import { ContentType } from "../domain/shared/content-type.vo";
 import { ETag } from "../domain/shared/etag.vo";
 import { ObjectKey } from "../domain/shared/object-key.vo";
-import { StoredObjectMetadata } from "../domain/stored-object/stored-object-metadata.entity";
+import { StoredObjectMetadata } from "../infra/stored-object/stored-object-metadata";
 import { SyncService } from "./sync.service";
 import type {
   IStoredObjectStorage,
   StoredObjectListItem,
 } from "../domain/shared/object-storage.interface";
-import type { IStoredObjectMetadataCommandRepository } from "../domain/stored-object/stored-object-metadata-command-repository.interface";
-import type { IStoredObjectMetadataQueryRepository } from "../domain/stored-object/stored-object-metadata-query-repository.interface";
+import type { IStoredObjectMetadataCommandRepository } from "../infra/stored-object/stored-object-metadata.command-repository.interface";
+import type { IStoredObjectMetadataQueryRepository } from "../infra/stored-object/stored-object-metadata.query-repository.interface";
 
 describe("SyncService", () => {
   describe("execute", () => {
